seperate config.py file
