--- kowalski/api/__init__.py
@@ -1,3 +1,0 @@
-from kowalski.api.api import *
-from kowalski.api.middlewares import *
-from kowalski.utils import *
