+++ kowalski/dask_clusters/dask_cluster_turbo.py
@@ -1,12 +1,12 @@
 import time
 
 from dask.distributed import LocalCluster
-from kowalski.utils import load_config, log
+from kowalski.utils import log
+from kowalski.config import load_config
 from kowalski.alert_brokers.alert_broker_turbo import WorkerInitializer  # noqa: F401
 
 
 """ load config and secrets """
-
 config = load_config(config_files=["config.yaml"])["kowalski"]
