+++ kowalski/ingesters/ingest_gaia_edr3.py
@@ -13,14 +13,13 @@ from tqdm.auto import tqdm
 from kowalski.utils import (
     deg2dms,
     deg2hms,
-    load_config,
     log,
     Mongo,
 )
+from kowalski.config import load_config
 
 
 """ load config and secrets """
-
 config = load_config(config_files=["config.yaml"])["kowalski"]
