+++ kowalski/ingesters/ingest_ps1_strm.py
@@ -13,13 +13,12 @@ from tqdm.auto import tqdm
 from kowalski.utils import (
     deg2dms,
     deg2hms,
-    load_config,
     log,
     Mongo,
 )
+from kowalski.config import load_config
 
 """ load config and secrets """
-
 config = load_config(config_files=["config.yaml"])["kowalski"]
