+++ kowalski/ingesters/ingest_ptf_matchfiles.py
@@ -18,14 +18,13 @@ from kowalski.utils import (
     deg2dms,
     deg2hms,
     init_db_sync,
-    load_config,
     log,
     Mongo,
 )
+from kowalski.config import load_config
 
 
 """ load config and secrets """
-
 config = load_config(config_files=["config.yaml"])["kowalski"]
 init_db_sync(config=config)
