+++ kowalski/tests/test_ingester_pgir.py
@@ -7,7 +7,8 @@ import requests
 from kowalski.alert_brokers.alert_broker_pgir import watchdog
 from kowalski.ingesters.ingester import KafkaStream
 from test_ingester_ztf import Program, Filter
-from kowalski.utils import Mongo, init_db_sync, load_config, log
+from kowalski.utils import Mongo, init_db_sync, log
+from kowalski.config import load_config
 
 """ load config and secrets """
