+++ kowalski/tools/fetch_ztf_matchfiles.py
@@ -11,7 +11,7 @@ from tqdm import tqdm
 from typing import Sequence
 from urllib.parse import urljoin
 
-from kowalski.utils import load_config
+from kowalski.config import load_config
 
 
 config = load_config(config_files=["config.yaml"])["kowalski"]
