+++ kowalski/tools/performance_reporter.py
@@ -12,10 +12,10 @@ import numpy as np
 from matplotlib.backends.backend_pdf import PdfPages
 from slack_sdk import WebClient
 from slack_sdk.errors import SlackApiError
-from kowalski.utils import load_config, log
+from kowalski.utils import log
+from kowalski.config import load_config
 
 """ load config and secrets """
-
 config = load_config(config_files=["config.yaml"])["kowalski"]
 
 action_patterns = {
