+++ requirements/requirements.txt
@@ -1,4 +1,3 @@
-bcrypt==4.0.1
 deepdiff==6.2.3
 fire==0.5.0
 gsutil==5.20
